PushBtnOpPro/app/requirements.txt
pyserial~=3.5
future~=0.18.2
setuptools